IR-685: Fix questions/answers case and add question marks to questions (#287)
* Added utility to add "?" to questions

A lot of the questions in NOMIS were clearly questions but they
lacked a "?". This utility simply check for some common words used
to ask questions (e.g. "who", "what", etc...) and if the label
is detected to be a question adds the question mark if it's not already
present.

This utility is conservative and it doesn't adds a "?" unless the label
is a question.

labels like `''ESTIMATED COST OF DAMAGE'` or `'DESCRIBE THE TYPE OF SPRAY'`
would be left alone.

However by inspection of several of the report configurations this should
correct most of the missing question marks.

* Utility to convert strings to sentence case

This function preserves the original case of some
acronyms or other special words (e.g. CCTV, F2052SH/ACCT
etc...)

More of these words could be added over time if needed.

* Fix case/question marks in questions labels

When converting NOMIS report types config questions'
labels are converted to sentence case and a question
mark is added if sentence is detected to be a question.

* Update case of reports config questions

* Updated Report charts after questions update

* Convert answers' case to sentence case (First pass)

* Sentence case: Added more preserve words

And updated reports configs.

* Updated reports config charts

* View report router uses `populateReport()`

The logic to get a report config is already in this
middleware.

* Moved `findAnswerConfigByCode()` alongside types

As it's now needed in the viewReport route as well.

* Report view: Use labels from config

Use questions/answers labels from the Report config
when displaying a report.

* Sentence case: Split by non-word boundaries

This solves a lot of issues were an acronym's case was
changed to lower case when not surrounded by space.
This lead to inconsistencies in the case of acronyms.
Sometimes even in the same sentence, e.g.
`Has the NOMS SFO team been informed (sfo cases only)?`

This is an improvement as a lot of the words we want
to preserve the case for are often surrounded by non-spaces
like `/` or parenthesis.

Unfortunately it's a tradeoff and this change means
some preserved words (e.g. `N/A`) which contains
some of the non-space characters are now not preserved.

However these words are not as frequent and usually
this doesn't look too bad and most importantly shouldn't
make them ambiguous. So overall this improves the case
of many of the questions/answers.

We could try and iterate further to try and improve
the case for these edge cases but it's possibly
excessive.

* Updated reports config charts

* Sentence case: Added more words as reserved

Few words suggested in code review: `GOAD`, `HMD`, `HMPPS`, `ORRU`
and `Vantive`.

Configs and charts updated accordingly.

* Fix question case fallback in `viewReport` route