PoC logic to determine questions to delete
Traverse the questionnaire tree to figure out which
of the questions/answers need to be retained.

This is already more useful than the basic "delete
all questions after current" without any additional
complexity. This means if a branch re-join the
flow of questions answered some of the future
questions/answers could still be preserved (as we
know they'll be asked anyway)

Example:
- Q1:
 - Yes -> Q2
 - No -> Q3
- Q2:
 - Yes/No -> Q3
- Q3:
 - ...

Here if I already answered Q1 (Yes), Q2 and Q3 and I
then I go back to Q1 and change the answer to "No"
I would still end up in Q3, so Q3's previous answer
can be preserved (Only Q2 needs to be deleted).

However the current implementation stops the traveral
the moment a question hasn't been answered yet.
This means that if a branch doesn't rejoin the main
flow future questions would still be deleted.
In other words, future questions are only preserved
if the questions/branch re-joins.

A possible way to prevent this is to check if any
sub-branching occurs, if not we can continue to
traverse the questionnaire even when the user hasn't
provided these answers (as they is no forking, they
will). This means that existing questions/answers
after the branch re-join could be preserved.

I have a rough idea of how to implement this
look-ahead logic but I don't know yet how much more
complacated it would make the traversal logic.