IR-732: Delete now-irrelevant questions when answering questions (#301)
* PoC logic to determine questions to delete

Traverse the questionnaire tree to figure out which
of the questions/answers need to be retained.

This is already more useful than the basic "delete
all questions after current" without any additional
complexity. This means if a branch re-join the
flow of questions answered some of the future
questions/answers could still be preserved (as we
know they'll be asked anyway)

Example:
- Q1:
 - Yes -> Q2
 - No -> Q3
- Q2:
 - Yes/No -> Q3
- Q3:
 - ...

Here if I already answered Q1 (Yes), Q2 and Q3 and I
then I go back to Q1 and change the answer to "No"
I would still end up in Q3, so Q3's previous answer
can be preserved (Only Q2 needs to be deleted).

However the current implementation stops the traveral
the moment a question hasn't been answered yet.
This means that if a branch doesn't rejoin the main
flow future questions would still be deleted.
In other words, future questions are only preserved
if the questions/branch re-joins.

A possible way to prevent this is to check if any
sub-branching occurs, if not we can continue to
traverse the questionnaire even when the user hasn't
provided these answers (as they is no forking, they
will). This means that existing questions/answers
after the branch re-join could be preserved.

I have a rough idea of how to implement this
look-ahead logic but I don't know yet how much more
complacated it would make the traversal logic.

* Questions deletion logic refactoring

Moved logic to determine questions to delete into
its own small service class with private helper
methods.

This should be a little bit easier to test.

* Fix existing Question router tests

Mock the requests to update the questions/answers
as the response is now used to determine which
questions will need to be deleted.

* Make requests to API to delete unreachable questions

On update make a request to the API to remove any previously answered
question which no longer applies (e.g. in a now unreachable branch)

* Tests for logic to determine questions to delete

* Made test data a bit more realistic.

Many of the test answers are answers which don't
require a date (`dateRequired: false` in config).
So the corresponsing responseDate would be `null`,
although it doesn't really change the flow of the
questions/answers or traversal in practice :)