IR-769: Validation on multiple choices questions
All answers should lead to the same next question. This is because
otherwise a user may choose answers in such a way that it is ambiguous
to know which question to show next.

Throughout the codebase we assume that when a question allow multiple choices
they all lead to the same next question (and just take the next question ID
from the 1st answer as it shouldn't matter).

Note that thix validation only considers active questions and active answers
as inactive ones don't matter.

Result of validation are in ticket: IR-793