PoC: Validate active config is now a test
This means committing an invalid configuration
will be flagged in a failing test locally or on CI.

I want to try and make this slightly more user-friendly
and ran a test for each individual config...but Jest'
`beforeAll()` is not cooperating.
In the meantime committing this as it is running the
validations even tho it would only show a failing test
for the first invalid configuration.