Made `getAllIncidentTypeConfigurations` synchronous
Trying to use this `async` utility function has been
a world of pain, especially in the context of Jest.

In practice it doesn't have to be `async`, it is only
used in the context of our validation tooling and
tests so we can live with it being synchronous/possibly
a bit slower if it means it's less of a pain to use.

More context: I was trying to use this function
to get the the incident report configurations and
have something along the line of (pseudocode):

```JavaScript
let allConfigs

beforeAll(async (done) => {
  allConfigs = await getAllIncidentTypeConfigurations()
  done()
})

it.each(allConfigs)('some name', config => {
  ...
})
```

But unfortunately Jest doesn't play ball, the `beforeAll()`
is not guaranteed to finish before the `it.each`
(or even a manual `for (const config of allConfigs)`)
so `allConfigs` is `undefined`. This was also the case
when I moving the `beforeAll()` in the parent `describe()`
block.