IR-970: Run incident type config validations as part of the test suite (#412)
* PoC: Validate active config is now a test

This means committing an invalid configuration
will be flagged in a failing test locally or on CI.

I want to try and make this slightly more user-friendly
and ran a test for each individual config...but Jest'
`beforeAll()` is not cooperating.
In the meantime committing this as it is running the
validations even tho it would only show a failing test
for the first invalid configuration.

* Made `getAllIncidentTypeConfigurations` synchronous

Trying to use this `async` utility function has been
a world of pain, especially in the context of Jest.

In practice it doesn't have to be `async`, it is only
used in the context of our validation tooling and
tests so we can live with it being synchronous/possibly
a bit slower if it means it's less of a pain to use.

More context: I was trying to use this function
to get the the incident report configurations and
have something along the line of (pseudocode):

```JavaScript
let allConfigs

beforeAll(async (done) => {
  allConfigs = await getAllIncidentTypeConfigurations()
  done()
})

it.each(allConfigs)('some name', config => {
  ...
})
```

But unfortunately Jest doesn't play ball, the `beforeAll()`
is not guaranteed to finish before the `it.each`
(or even a manual `for (const config of allConfigs)`)
so `allConfigs` is `undefined`. This was also the case
when I moving the `beforeAll()` in the parent `describe()`
block.

* Validate incident types configs in separate tests

Use `it.each` to run the validation on the various
active incident type configuration separately, so that
the output is cleaner.

Example run:

```
 FAIL  server/data/incidentTypeConfiguration/validation.test.ts
  Active incident type configurations
    ✓ Config for ABSCOND_1 is valid (1 ms)
    ✓ Config for ASSAULT_5 is valid
    ✓ Config for ATTEMPTED_ESCAPE_FROM_PRISON_1 is valid
    ✓ Config for ATTEMPTED_ESCAPE_FROM_ESCORT_1 is valid
    ✓ Config for BOMB_1 is valid
    ✓ Config for BREACH_OF_SECURITY_1 is valid (1 ms)
    ✓ Config for CLOSE_DOWN_SEARCH_1 is valid
    ✓ Config for DEATH_PRISONER_1 is valid
    ✓ Config for DEATH_OTHER_1 is valid
    ✓ Config for DISORDER_2 is valid
    ✓ Config for DRONE_SIGHTING_3 is valid
    ✓ Config for ESCAPE_FROM_PRISON_1 is valid
    ✓ Config for ESCAPE_FROM_ESCORT_1 is valid (1 ms)
    ✕ Config for FIND_6 is valid
    ✓ Config for FIRE_1 is valid (1 ms)
    ✓ Config for FOOD_REFUSAL_1 is valid
    ✓ Config for KEY_OR_LOCK_2 is valid
    ✓ Config for MISCELLANEOUS_1 is valid
    ✓ Config for RADIO_COMPROMISE_1 is valid
    ✓ Config for RELEASE_IN_ERROR_1 is valid
    ✓ Config for SELF_HARM_1 is valid (1 ms)
    ✓ Config for TEMPORARY_RELEASE_FAILURE_4 is valid
    ✓ Config for TOOL_LOSS_1 is valid

[...]

  ● Active incident type configurations › Config for FIND_6 is valid

    Config for 'FIND_6' incident type is invalid: the following multiple choices questions can lead to different next questions: 67179; some answers lead to these unknown or inactive questions: 67180-BREAKME

      13 |     const errors = validateConfig(config).map(err => err.message)
      14 |     if (errors.length > 0) {
    > 15 |       throw new Error(
         |             ^
      16 |         `Config for '${incidentType}' incident type is invalid: ${errors.join('; ')}`,
      17 |       )
      18 |     }

      at server/data/incidentTypeConfiguration/validation.test.ts:15:13

Test Suites: 1 failed, 1 total
Tests:       1 failed, 35 passed, 36 total
Snapshots:   0 total
Time:        0.64 s, estimated 1 s
```

* Appease linter

* Validate that questions/answers IDs don't include hyphens

There are parts of the codebase that now split form field
names by `-`, so while IDs are still strings we want
to be sure they don't include an hyphen as this can
lead to problems.

See: https://github.com/ministryofjustice/hmpps-incident-reporting/pull/403#discussion_r2028814884