Use HMPPS library's RestClient for FrontendComponentsClient
The `FrontendComponentsClient` is a bit special in the
sense that it doesn't extend the `RestClient` but
rather instantiates one when making the request to get
the components.

But `hmpps-rest-client`'s `RestClient` is now an abstract
class, so it can't be instantiated directly, hence
the introduction of the little wrapper class `ConcreteRestClient`
which only serves the purpose of being able to instantiate
a `RestClient`.

NOTE: AFAICT We can't extend `FrontendComponentsClient`
from `RestClient` directly because this is instanciated
at app start up rather than on a per-request basis.