IR-1027: Backport template changes (#431)
* Change deprecated tsconfig option (#534)

* Fixing error handling (#540)

* SDIT-2709: 🔥 Remove config classes as now in typescript lib (#542)

* Use HMPPS library's `RestClient` for IncidentReportingApi

* Use HMPPS library's RestClient for NomisUserRolesApi

* Use HMPPS library's RestClient for OffenderSearchApi

* Use HMPPS library's RestClient for PrisonApi

* Fix authentication of IncidentReportingApi

* Fixed authentication of NomisUserRolesApi

* Fixed authentication of OffenderSearchApi

* Use HMPPS library's RestClient for FrontendComponentsClient

The `FrontendComponentsClient` is a bit special in the
sense that it doesn't extend the `RestClient` but
rather instantiates one when making the request to get
the components.

But `hmpps-rest-client`'s `RestClient` is now an abstract
class, so it can't be instantiated directly, hence
the introduction of the little wrapper class `ConcreteRestClient`
which only serves the purpose of being able to instantiate
a `RestClient`.

NOTE: AFAICT We can't extend `FrontendComponentsClient`
from `RestClient` directly because this is instanciated
at app start up rather than on a per-request basis.

* Use HMPPS library's `RestClient` for `ManageUsersApiClient`

`ManageUsersApiClient` is a bit special in the sense
that it doesn't extend `RestClient` but rather it
instantiances it which is why it's now using the
`ConcreteRest` wrapper class (as `RestClient` is now
an abstract class and can't be instantianted)

* Removed unused usage of RestClient in HmppsAuthClient

Unit and integration tests are passing and it doesn't seem
to be used.

* Removed now unsed local `RestClient` class

This is now provided by the `hmpps-rest-client`
library.

* Use token store classes/interfaces from hmpps-auth-clients

* Use `hmpps-auth-clients`' `VerificationClient`

for token verification

* Use `hmpps-auth-clients`' AuthenticationClient

* Use `SanitisedError` in HMPPS library

* Use latest version of @ministryofjustice/hmpps-auth-clients

* Revert assertion to concrete error classes

See PR comment: https://github.com/ministryofjustice/hmpps-incident-reporting/pull/431#discussion_r2084451387

---------

Co-authored-by: Paul Solecki <51918433+psoleckimoj@users.noreply.github.com>
Co-authored-by: Andrew Lee <1517745+andrewrlee@users.noreply.github.com>
Co-authored-by: petergphillips <5099053+petergphillips@users.noreply.github.com>