Refactored worklist statuses logic into a function
The `incidentStatuses` value coming from the query param could be
`undefined` when not provided, it could be a string when only a checkbox
is checked, it could be a list of strings which are either Status codes
or worklist code (e.g. `toDo`, `submitted` or `done`).

This value has to be massaged to:
- turned into a list of strings if it's a string
- check for invalid statuses/checklist codes
- turn into a list of Status codes to be passed to the API for filtering

There are 2 branches, depending on whether the user is a Data Warden (they
see all the statuses) or a Reporting Officer (they see high-level "worklists")

Most of this logic is now inside an utility function rather than directly
being scatterred across the request handler.