IR-1062: Statuses checkboxes display bug (#449)
* govukMultipleCheckedItems() throws exception...

...when argument which is supposed to be an array of strings
is not. This is to prevent a bug like IR-1062 to happen again.

This uses `includes()` which is defined on both
arrays and strings...when a string is passed as an
argument the function would "check" options that
it shouldn't because they're substrings of the
selected option.

* Dashboard: When provided, `searchStatuses` is an array

This is to avoid bugs where potentially a checkbox for a status is checked
(e.g. `CLOSED`) when it shouldn't (because only `WAS_CLOSED` was selected)

* Tweaks to type annotations

* Refactored worklist statuses logic into a function

The `incidentStatuses` value coming from the query param could be
`undefined` when not provided, it could be a string when only a checkbox
is checked, it could be a list of strings which are either Status codes
or worklist code (e.g. `toDo`, `submitted` or `done`).

This value has to be massaged to:
- turned into a list of strings if it's a string
- check for invalid statuses/checklist codes
- turn into a list of Status codes to be passed to the API for filtering

There are 2 branches, depending on whether the user is a Data Warden (they
see all the statuses) or a Reporting Officer (they see high-level "worklists")

Most of this logic is now inside an utility function rather than directly
being scatterred across the request handler.

* Refactoring: Renamed `checkForOutliers()` => `hasInvalidValues()`

This is what this function is really doing: Checking if a value or list
of values is valid or not against a list of allowed values.