Use GitHub actions as CI (#459)
* Copied GH Actions workflow from TS template

* Explicly run tests in Europe/London timezone

So that they don't fail in CI

* Use docker image in GH container registry

* Updated READMEs to replace CircleCI

Added badge linking to ghcr.io Docker images and other references to
CircleCI.

* Point to new application-insights secret

This is now generated by the `cloud-platform-terraform-hmpps-template`
terraform module.

* Removed CircleCI workflows

* Added deploy to preprod/prod jobs

NOTE: They would not work until CP environment changes to add these
enviroments to the GH repositories (and populate the deploy secrets)
are applied.

* Set timezone when running integration tests interactively

For consistency with the other scripts in the `packages.json`.
Although we may find a better way of setting the timezone for CI separately
at some point.