IR-1047: Categorise users into types before determining permissions (#463)
This abstraction makes it easier to reason about the logic of each permission check and to compare against designs and requirements.

Previously, several checks required the presence of `roleReadWrite` _and the absence_ of `roleApproveReject`. That’s because `roleReadWrite` flagged a reporting officer and `roleApproveReject` flagged a data warden… but a user cannot be both!

Now this is just `isReportingOfficer` because the most powerful role determines the user type and all allowable actions stem from there.

---

Also, mock user variables have been renamed to follow these user type groupings.