Renamed `findAnswerConfigByCode()` to `findAnswerConfigByResponse()`
As the field in the answer configu is now called `response`, this is
the text identifier for an answer, e.g. `'YES'` rather than the NOMIS
numeric one (e.g. `12345`).

Various variables have been renamed as a result and the `findAnswerConfigForResponse()`
utility function which was just wrapping this utility function by calling
`findAnswerConfigByResponse(response.response, ...` has been removed and inlined.