IR-1097: Renaming config fields to match API names (#474)
* Renamed QuestionConfiguration's `code` to `question`

This is to match field name in the API.

* Renamed AnswerConfiguration's `code` to `response`

This is to match field name in the API.

* Renamed question `id` into `code`

This is the numeric code (e.g. `'12345'`), not the string
(e.g. `'WHY?'`)

* Renamed answers' id to code

To be consistent with API. Note the `code` here
is the numeric code (e.g. `'12345'`) not the constant
string for the answer (e.g. `'YES'`, in the API this is
named `response`)

* Renamed `startingQuestionId`/`nextQuestionId` to `*code`

* Rename fields in DOT/graph nunjucks template

* Don't swallow `spawnSync()` errors

Utilities to generate Report types config and
graphs use `spawnSync()` but they didn't report any
potential error of the child process. Errors are now
thrown.

Also moved typescript formatting logic into a small
utility function for reuse across the couple of
places where we do that.

NOTE: The `stdio` option of `spawnSync()` already
defaults to `'pipe'` so I've removed that as it's
just noise, see: https://nodejs.org/api/child_process.html#child_processspawnsynccommand-args-options

* linting

* Removed unnecessary `questionFieldName()` util

This existest because of fields being called differently
in UI and API (`code` was called `id` in the UI report
configuration).

Now question have `code` both on the API and on the
UI side.

* Tweaked various utility functions to use question/response "code"

Not "id".

* Renamed `findAnswerConfigByCode()` to `findAnswerConfigByResponse()`

As the field in the answer configu is now called `response`, this is
the text identifier for an answer, e.g. `'YES'` rather than the NOMIS
numeric one (e.g. `12345`).

Various variables have been renamed as a result and the `findAnswerConfigForResponse()`
utility function which was just wrapping this utility function by calling
`findAnswerConfigByResponse(response.response, ...` has been removed and inlined.

* Renamed more id/code/response references

* Corrected spelling of hyphen