Reworked Permissions middleware to use `getServicePrisonIds()`
The `activePrisons` is no longer read from the config and injected
in the Permission class and other methods needing to check if a prison
is active.

NOTE: Many places still need updating, for example the `isLocationActiveInService()`
utility function is assumed to get the list of active prisons as a parameter
now but it needs to be updated.