Middleware to update the list of active prisons
This middleware gets a fresh (possibly cached)
version of the list of prisons where the service is
active and both sets `res.locals.activePrisons` but
also updates the `applicationInfo` singleton.

The latter is important as the `/info` endpoint
exposes the list of active prisons via the `applicationInfo`
object: By updating it the `/info` endpoint would
automatically get an up-to-date version of the
active prisons (without needing to restart the
application for example).