Use new Prison API `/api/agency-switches/` endpoint
Beside the endpoint path the response is slightly different as it
now refers to "agencies" rather than "prisons" (as it can switch a service
on non-prisons, e.g. PECS regions in our case).

As a result I've also renamed various references to "prisons" to "agencies".
Maily the `activePrisons` in res.locals and various argument names and
variables.