Moved `activeAgencies` cache to `updateActiveAgencies`
This means that potentially less requests to HMPPS Auth (or rather its
TokenStore cache) are made and also that we don't update `activeAgencies`
unneccarily with the same "new" value.

As a result of this the cache now doesn't actually contains the list
of active agencies (as we're not memoising `getAgenciesSwitchedOn()`)
but it is simply a way to check when the list was updated last (and
leaving the `activeAgencies` list untouched if it was updated recently)

Admittedly the use of Symbol is possibly overkill? A simple value like a
boolean would have sufficed but a symbol is possibly more explicit than
something like `true`. Although the extra type required by TypeScript makes
this more cumbersome that it probably should.