updateActiveAgencies: Handle errors gracefully
The assumption being that if sporadic errors to
get an updated list happen we can continue to use
the existing copy of the active agencies without
disruprive requests.

Main drawback with this is that if the application
starts and never manage to get a good copy of the
active agencies and the Prison API is consistently
broken the exception is only logged and could go
unnoticed.