IR-1117: Get active agencies from Prison API's `/api/agency-switches/INCIDENTS` endpoint (#477)
* Prison API: Method to get prisons where service is active

Uses [`GET /api/service-prisons/{serviceCode}` endpoint] to
get the list of prisons where the service (code: `INCIDENTS`)
is active. The idea is to replace the current
`config.activePrisons` feature flag.

[`GET /api/service-prisons/{serviceCode}` endpoint]: https://prison-api-dev.prison.service.justice.gov.uk/swagger-ui/index.html#/service-prisons/getServicePrisons

* Prison API: Service prisons method returns just ids

* Export special prison ID `'*ALL*'`

This is what `/api/service-prisons/{serviceCode}`
uses to signify the given service is on in all prisons.

* `isLocationActiveInService()` to use argument list of prison Ids

These functions now check if the service is active in
a prison by checking its id against the provided
`activePrisons` argument rather than using
`config.activePrisons`

* Reworked Permissions middleware to use `getServicePrisonIds()`

The `activePrisons` is no longer read from the config and injected
in the Permission class and other methods needing to check if a prison
is active.

NOTE: Many places still need updating, for example the `isLocationActiveInService()`
utility function is assumed to get the list of active prisons as a parameter
now but it needs to be updated.

* Fix nunjucksSetup test

* Fix the bulk of the failing tests

The issue was the same in all of these: Was trying to get the list
of prisons where the service is active from Prison API and the request
needs to be mocked.

* Revert "Fix the bulk of the failing tests"

This reverts commit 113d203167c66fb53581ca74d01cf73ea7de0cc0.

* Fix test for logoutUnless

* Revert "Fix test for logoutUnless"

This reverts commit e664788b014ee654218413081a8a7aa67844b704.

* Removed unused import in a test

* Middleware to update the list of active prisons

This middleware gets a fresh (possibly cached)
version of the list of prisons where the service is
active and both sets `res.locals.activePrisons` but
also updates the `applicationInfo` singleton.

The latter is important as the `/info` endpoint
exposes the list of active prisons via the `applicationInfo`
object: By updating it the `/info` endpoint would
automatically get an up-to-date version of the
active prisons (without needing to restart the
application for example).

* Permission middleware to use res.locals.activePrisons

* Added `activePrisons` to Response type

* Fix tests using `res.locals.activePrisons`

* Cache getServicePrisonIds() result for 1 hour

Added very simple in-memory caching mechanism to
avoid making unnecessary requests to `GET /api/service-prisons/INCIDENTS`

I may move the logic in a generic Cache class

* Extracted simple caching logic into Cache<T> class

* Removed `config.activePrisons`/`SERVICE_ACTIVE_PRISONS`

Unused now that application gets this information from
`GET /api/service-prisons/INCIDENTS`

* Removed unused `config` import in a test

* REMOVE: DEBUG: Logging on cache/cache for 30s

TODO: Remove commit.

* removed `isPrisonActiveInService()`

Using `isLocationActiveInService()` for both prisons
and PECS region. The mechanism to determine whether
service is active or not is the same: Making a
request to the Prison API's "agency-switches" endpoint
(was `GET /api/service-prisons/INCIDENTS`)

* Removed `config.activeForPecsRegions`/`SERVICE_ACTIVE_PECS` flag

Service is activated for PECS regions in the same way as for
other locations: Via the Prison API `/api/agency-switches/`
endpoint

* Use new Prison API `/api/agency-switches/` endpoint

Beside the endpoint path the response is slightly different as it
now refers to "agencies" rather than "prisons" (as it can switch a service
on non-prisons, e.g. PECS regions in our case).

As a result I've also renamed various references to "prisons" to "agencies".
Maily the `activePrisons` in res.locals and various argument names and
variables.

* Revert "REMOVE: DEBUG: Logging on cache/cache for 30s"

This reverts commit 017ce9a77cd26584f42635a2d959b564b6752dc7.

* Fix integration tests

* Linting: Removed unused import

* Renamed `ServiceAgency` to `ActiveAgency`

It's a better name for the items returned by the "agency-switches"
endpoint.

(and we refer to these as "active agencies" across the codebase now)

* Moved `SERVICE_ALL_AGENCIES` in `activeAgencies` module

This module will have more stuff in it soon.

* Renamed `updateActiveAgencies`

Was still called `updateActivePrisons`

* activeAgencies is now a singleton

This avoid the need to pass `activeAgencies` as a
parameter which is particularly cumbersome in nunjucks
templates.

The rest is mostly the same but instead of getting this
from `res.locals.activeAgencies` this is now a singleton
in its own module.

* Updated tests to use new `activeAgencies` singleton

* Improved comment in applicationInfo explaing who updates activeAgencies

* Released requirement for `'*ALL*'` to be only active agency

`isLocationActiveInService()` now doesn't care if there are other agency
ids alongside the special `'*ALL*'` marker, it would still return
always true in that case.

* Added a test for updateActiveAgencies middleware

* updateActiveAgencies: Improved types and documentation

* Fix "all agencies" in DPS' application info activeAgencies

DPS interprets the `applicationInfo.additionalFields.activeAgencies`
slightly differently in regards to "active in all services":
- Prison API uses `'*ALL*'` to signify all...
- ...DPS uses `'***'`

This is to be sure that applicationInfo uses the DPS convention

* Moved `activeAgencies` cache to `updateActiveAgencies`

This means that potentially less requests to HMPPS Auth (or rather its
TokenStore cache) are made and also that we don't update `activeAgencies`
unneccarily with the same "new" value.

As a result of this the cache now doesn't actually contains the list
of active agencies (as we're not memoising `getAgenciesSwitchedOn()`)
but it is simply a way to check when the list was updated last (and
leaving the `activeAgencies` list untouched if it was updated recently)

Admittedly the use of Symbol is possibly overkill? A simple value like a
boolean would have sufficed but a symbol is possibly more explicit than
something like `true`. Although the extra type required by TypeScript makes
this more cumbersome that it probably should.

* Split `updateActiveAgencies` middleware tests

This is to test different cases separately

* Shorten `activeAgencies` when containing `'*ALL*'`

We update infrequently because of the cache but the
look-up is done several times: When the `activeAgencies`
list contains `'*ALL*'` all other potential active
agencies are removed as `'*ALL*'` includes everything
anyway.

* updateActiveAgencies: Handle errors gracefully

The assumption being that if sporadic errors to
get an updated list happen we can continue to use
the existing copy of the active agencies without
disruprive requests.

Main drawback with this is that if the application
starts and never manage to get a good copy of the
active agencies and the Prison API is consistently
broken the exception is only logged and could go
unnoticed.

* Added `nock.clearAll()` in updateActiveAgencies tests

To be sure that HMPPS Auth request is not mocked
multiple times.

* Made updateActiveAgencies test less reliant on order

The scenario where the request to update activeAgencies
was relying on a previous test to have ran. This change
ensure the state is well-defined at the beginning
of the test.

* updateActiveAgencies test: removed unnecessary `satisfies`