IR-1134: Handle 404 from /agencies-switches (#479)
In productions we are seeing 404 Not Found errors.
These are handled ok but the interesting thing is
that this is the way that endpoing behaves when no
agencies are active.

This commit handle this scenario more explicitly so
that the errors are not concerning anybody and also
to document this is the way the endpoint works.