IR-1363: Adopted `hmpps-npm-script-allowlist` (#649)
* Basic setup of hmpps-npm-script-run-allowlist

This will prevent scripts from running unless they're explicitly allowed
to do so.

* Updated list of npm scripts allowed to run

Most of the comments were taken from this PR in the TS template:
https://github.com/ministryofjustice/hmpps-template-typescript/pull/632/files#diff-be0b106dba7c5a3635614a4e23e4e29c9da4462922769c8f99b60e31d288c7b1

* Install hmpps-npm-script-allowlist dev dependency

This is the HMPPS package responsible for allowing some of the npm scripts
to run (allow list is in the `.allowed-scripts.mjs` file).

* Security bumps: Upgrade `body-parser` and `jws`

Both packages were vulnerable:

- `body-parser` `2.2.0`: https://github.com/expressjs/body-parser/security/advisories/GHSA-wqch-xfxh-vrr4
- `jws` `3.2.2`: https://github.com/advisories/GHSA-869p-cjfg-cm3x

Updated them using `npm audit fix`.