IR-1354: Optional comments Proof of Concept
Background
----------
In NOMIS answers' comments are optional, that means that even when the answer
configuration "requires" it, the user can submit a report without providing one.

In DPS when the "comment required" configuration is set we interpret this as
"comment is mandatory".

Field renaming
--------------
We're introducing a new field called `commentRequested` which mean user will
be shown a comment input for that answer.

Whether that comment is mandatory (default/existing behaviour) or optional
depends on the `commentMandatory` field.

This means:
- the old `commentRequired = false` (meaning the comment will not be shown)
  is replaced with `commentRequested = false`/`commentMandatory = false`
- the old `commentRequired = true` (meaning the comment will be shown and it's
  mandatory) is replaced with
  `commentRequested = true`/`commentMandatory = true`
- in future we may have some answers configured to optionally ask for a
  comment, that is the user can provide a comment if they want to but it's
  not mandatory. This will be achieved with
  `commentRequested = true`/`commentMandatory = false`
- the `commentRequested = false`/`commentMandatory = true` is clearly **invalid** because it would mean a comment is mandatory but the user is not
  prompted to enter one (so he cannot)
  - this should be checked with a new validation